pyproject.toml
[tool.poetry]
name = "kwja"
version = "0.1.0"
description = "A unified language analyzer for Japanese"
authors = [
    "Hirokazu Kiyomaru <kiyomaru@i.kyoto-u.ac.jp>",
    "Nobuhiro Ueda <ueda@nlp.i.kyoto-u.ac.jp>",
    "Kazumasa Omura <omura@nlp.i.kyoto-u.ac.jp>",
    "Takashi Kodama <kodama@nlp.i.kyoto-u.ac.jp>",
    "Yugo Murawaki <murawaki@i.kyoto-u.ac.jp>",
    "Daisuke Kawahara <dkw@waseda.jp>",
    "Sadao Kurohashi <kuro@i.kyoto-u.ac.jp>"
]
keywords = ["natural language processing", "japanese"]

packages = [
    { include = "kwja", from = "src" }
]
include = [
    "kwja/resource/**/*"
]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
torch = "~1.11.0"
pytorch-lightning = "^1.6.5"
omegaconf = "^2.1"
hydra-core = "^1.2"
tokenizers = "^0.12.1"
transformers = "~4.19.2"
typer = "^0.6.1"
sentencepiece = "^0.1.96"
protobuf = "^3.20"
seqeval = "^1.2.2"
wandb = "^0.13.3"
rich = "^12.4.4"
rhoknp = "^0.4.0"
python-Levenshtein = "^0.12.2"
fugashi = "^1.1.2"
ipadic = "^1.0.0"
python-dotenv = "^0.20.0"
pandas = "^1.4.2"
dartsclone = "^0.10.2"
kyoto-reader = "^2.5.2"
jinf = "^1.0.4"
tinydb = "^4.7.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1"
pytest-cov = "^3.0.0"

[tool.poetry.scripts]
kwja = "kwja.cli.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py39", "py310"]

[tool.isort]
line_length = 120
py_version = 39
profile = "black"
src_paths = ["src"]

[tool.mypy]
python_version = 3.9
